Also check for alias in the logger compiler pass
The logger compiler pass only checks the container for a service whose
id is 'logger', but for version <=3.0.0 of Monolog, 'logger' is only an
alias, not an id. Because of this, the logger used by the Gentle Force
bundle defaults to NullLogger and logging fails to happen.

With this change, we check both for a service whose id is 'logger' OR
for a service whose alias is 'logger'. This makes it compatible with
versions of Monolog below 3.1.x.